pyproject.toml
[tool.poetry]
name = "replace_pics"
version = "0.1.0"
description = ""
authors = ["bedlavedlt <bedlavedlt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
